+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## [Unreleased]
 
-### Fixed
+### Fixed 
 
+- Fix drilldown regroup strategy on fake-split charts.
 - From now vertical line connections are curved lines.
 - Remove duplicated circles on line-circle transition.
 - Fix area-circle polar connection transition.